+++ script/convert_miopen_driver_to_profiler.py
@@ -31,10 +31,9 @@ def parse_layouts(args):
        args.in_layout == "NCDHW":
         if args.ck_profier_op == "grouped_conv_bwd_weight":
             args.layout = 4
-        elif args.ck_profier_op == "grouped_conv_fwd":
+        elif args.ck_profier_op == "grouped_conv_fwd" or \
+             args.ck_profier_op == "grouped_conv_bwd_data":
             args.layout = 3
-        elif args.ck_profier_op == "grouped_conv_bwd_data":
-            args.layout = 2
         else:
             print('Not supported layout for this op')
             exit(1)